Fold ConfigurationBootstrapper in AddCollectionConfigurations
