--- src/Workleap.Extensions.Mongo.Abstractions/IMongoCollectionConfigurationBootstrapper.cs
@@ -1,8 +1,0 @@
-using System.Reflection;
-
-namespace Workleap.Extensions.Mongo;
-
-public interface IMongoCollectionConfigurationBootstrapper
-{
-    void ApplyConfigurations(params Assembly[] assemblies);
-}
