+++ src/Workleap.Extensions.Mongo.Abstractions/MongoReflectionCache.cs
@@ -27,6 +27,16 @@ internal static class MongoReflectionCache
     internal static bool IsConcreteMongoDocumentType(this Type type) => !type.IsAbstract && typeof(IMongoDocument).IsAssignableFrom(type);
     
     internal static bool IsMongoCollectionConfigurationInterface(this Type t) => t.IsGenericType && t.GetGenericTypeDefinition() == typeof(IMongoCollectionConfiguration<>);
+    
+    internal static bool HasPublicParameterlessConstructor(this Type type) => type.GetConstructor(BindingFlags.Public | BindingFlags.Instance, binder: null, Type.EmptyTypes, modifiers: null) != null;
+
+    internal static void EnsureHasPublicParameterlessConstructor(this Type type)
+    {
+        if (!type.HasPublicParameterlessConstructor())
+        {
+            throw new InvalidOperationException($"Type {type}' must have a public parameterless constructor");
+        }
+    }
 }
 
 internal sealed class MongoReflectionCacheConfigurationStrategy : IMongoReflectionCacheStrategy
